+++ sshd-core/src/main/java/org/apache/sshd/common/session/Session.java
@@ -67,6 +67,14 @@ public interface Session
                 Closeable {
 
     /**
+     * Quick indication if this is a server or client session (instead of
+     * having to ask {@code instanceof}).
+     *
+     * @return {@code true} if this is a server session
+     */
+    boolean isServerSession();
+
+    /**
      * Timeout status.
      */
     enum TimeoutStatus {